jenkins file updated

+++ sshConnect.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
-sudo curl -o /var/www/html/index.html https://raw.githubusercontent.com/Andrey15716/DOS15-onl_DevOps_exam/main/application/index.html
-sudo systemctl restart nginx
+
+if curl -f -s -o /dev/null https://raw.githubusercontent.com/Andrey15716/DOS15-onl_DevOps_exam/main/application/index.html; then
+  sudo curl -o /var/www/html/index.html https://raw.githubusercontent.com/Andrey15716/DOS15-onl_DevOps_exam/main/application/index.html
+  sudo systemctl restart nginx
+else
+  echo "Приложение недоступно"
+  exit 1
+fi
